fixed spec tests

+++ spec/stackbuilder/chef/chef_stack_spec.rb
@@ -6,6 +6,18 @@ include StackBuilder::Common::Helpers
 
 describe StackBuilder::Chef do
 
+    # Change the following to true if you want to create
+    # the repo in your ~/tmp folder for debugging
+    INTERACTIVE_DEBUG_MODE = false
+
+    # Comment the tests below to make the test
+    # run shorter when debugging issues
+    RUN_TESTS = [
+        :wordpress,
+        :generic,
+        :containter
+    ]
+
     before(:all) do
 
         @logger = StackBuilder::Common::Config.logger
@@ -14,7 +26,7 @@ describe StackBuilder::Chef do
 
         @test_data_path = File.expand_path('../../../data', __FILE__)
 
-        @tmp_dir = Dir.mktmpdir
+        @tmp_dir = INTERACTIVE_DEBUG_MODE ? Dir.home + '/tmp' : Dir.mktmpdir
         @repo_path = "#{@tmp_dir}/test_repo"
 
         @stack_ids = { :dev => 'DEV', :test => 'TEST', :prod => 'PROD' }
@@ -36,32 +48,9 @@ describe StackBuilder::Chef do
     end
 
     after(:all) do
-        FileUtils.rm_rf(@tmp_dir)
+        FileUtils.rm_rf(@tmp_dir) unless INTERACTIVE_DEBUG_MODE
 
         Dir.chdir(@cwd)
-    end
-
-    it "should upload the stack" do
-
-        knife_cmd = Chef::Knife::StackInitializeRepo.new
-        knife_cmd.config[:repo_path] = @repo_path
-        knife_cmd.config[:certs] = 'wpweb.stackbuilder.org,wpdb.stackbuilder.org'
-        knife_cmd.config[:stack_environments] = 'DEV,TEST,PROD'
-        knife_cmd.config[:cookbooks] =
-            'haproxy:=1.6.6,' +
-            'mysql:=5.6.1,' +
-            'mysql-chef_gem:=0.0.5,' +
-            'apache2:=2.0.0,' +
-            'wordpress:=2.3.0'
-
-        knife_cmd.run
-
-        # Copy the test data into the repo
-        system("cp -fr #{@test_data_path}/test_repo/* #{@repo_path}")
-
-        knife_cmd = Chef::Knife::StackUploadRepo.new
-        knife_cmd.config[:repo_path] = @repo_path
-        run_knife(knife_cmd)
     end
 
     def validate_nodes(nodes, container_node = nil)
@@ -115,12 +104,12 @@ describe StackBuilder::Chef do
 
         nodes.each do |n|
 
-            next if n.start_with?('database')
-
             node = Chef::Node.load(n)
             container_port = container_port_map[node['fqdn']]
 
-            if container_port.nil?
+            if container_node.nil?
+
+                next if n.start_with?('database')
 
                 ip = node['ipaddress']
 
@@ -128,7 +117,9 @@ describe StackBuilder::Chef do
                 expect(resp.code).to eq('200')
 
                 wordpress_ips << ip if n.start_with?('wordpress')
-            else
+
+            elsif !container_port.nil?
+
                 timeout(30) do
                     resp = http_fetch("http://#{container_node['ipaddress']}:#{container_port}")
                     expect(resp.code).to eq('200')
@@ -151,160 +142,188 @@ describe StackBuilder::Chef do
         end
     end
 
-    it "should build a DEV stack" do
+    it "should test stack build, scale and destroy" do
 
-        knife_cmd = Chef::Knife::StackBuild.new
-        knife_cmd.name_args = [ @test_data_path + '/test_repo/stack1.yml' ]
-        knife_cmd.config[:stack_id] = @stack_ids[:dev]
-        knife_cmd.config[:environment] = 'DEV'
+        knife_cmd = Chef::Knife::StackInitializeRepo.new
         knife_cmd.config[:repo_path] = @repo_path
+        knife_cmd.config[:certs] = 'wpweb.stackbuilder.org,wpdb.stackbuilder.org'
+        knife_cmd.config[:stack_environments] = 'DEV,TEST,PROD'
+        knife_cmd.config[:cookbooks] =
+                'haproxy:=1.6.6,' +
+                'mysql:=5.6.1,' +
+                'mysql-chef_gem:=0.0.5,' +
+                'apache2:=2.0.0,' +
+                'wordpress:=2.3.0'
+
         knife_cmd.run
 
-        validate_nodes(['database-DEV-0', 'wordpress-DEV-0', 'loadbalancer-DEV-0'])
-    end
+        # Copy the test data into the repo
+        system("cp -fr #{@test_data_path}/test_repo/* #{@repo_path}")
 
-    it "should scale up the DEV stack web tier" do
-
-        knife_cmd = Chef::Knife::StackBuild.new
-        knife_cmd.name_args = [ @test_data_path + '/test_repo/stack1.yml' ]
-        knife_cmd.config[:stack_id] = @stack_ids[:dev]
-        knife_cmd.config[:environment] = 'DEV'
-        knife_cmd.config[:node] = 'wordpress:2'
+        knife_cmd = Chef::Knife::StackUploadRepo.new
         knife_cmd.config[:repo_path] = @repo_path
-        knife_cmd.run
+        run_knife(knife_cmd)
 
-        validate_nodes(['database-DEV-0', 'wordpress-DEV-0', 'wordpress-DEV-1', 'loadbalancer-DEV-0'])
+        if RUN_TESTS.include?(:wordpress)
+
+            puts "Building DEV stack"
+
+            knife_cmd = Chef::Knife::StackBuild.new
+            knife_cmd.name_args = [ @test_data_path + '/test_repo/stack1.yml' ]
+            knife_cmd.config[:stack_id] = @stack_ids[:dev]
+            knife_cmd.config[:environment] = 'DEV'
+            knife_cmd.config[:repo_path] = @repo_path
+            knife_cmd.run
+
+            validate_nodes(['database-DEV-0', 'wordpress-DEV-0', 'loadbalancer-DEV-0'])
+
+            puts "Scaling up the DEV stack web tier"
+
+            knife_cmd = Chef::Knife::StackBuild.new
+            knife_cmd.name_args = [ @test_data_path + '/test_repo/stack1.yml' ]
+            knife_cmd.config[:stack_id] = @stack_ids[:dev]
+            knife_cmd.config[:environment] = 'DEV'
+            knife_cmd.config[:node] = 'wordpress:2'
+            knife_cmd.config[:repo_path] = @repo_path
+            knife_cmd.run
+
+            validate_nodes(['database-DEV-0', 'wordpress-DEV-0', 'wordpress-DEV-1', 'loadbalancer-DEV-0'])
+
+            puts "Scaling down the DEV stack web tier"
+
+            knife_cmd = Chef::Knife::StackBuild.new
+            knife_cmd.name_args = [ @test_data_path + '/test_repo/stack1.yml' ]
+            knife_cmd.config[:stack_id] = @stack_ids[:dev]
+            knife_cmd.config[:environment] = 'DEV'
+            knife_cmd.config[:node] = 'wordpress:1'
+            knife_cmd.config[:repo_path] = @repo_path
+            knife_cmd.run
+
+            validate_nodes(['database-DEV-0', 'wordpress-DEV-0', 'loadbalancer-DEV-0'])
+
+            puts "Destroying the DEV stack"
+
+            stack = StackBuilder::Stack::Stack.new(
+                StackBuilder::Chef::NodeProvider.new(@repo_path, 'DEV'),
+                @test_data_path + '/test_repo/stack1.yml',
+                @stack_ids[:dev] )
+
+            stack.destroy
+
+            knife_cmd = Chef::Knife::NodeList.new
+            node_list = run_knife(knife_cmd).split
+            expect(node_list.empty?).to be_truthy
+        end
+
+        if RUN_TESTS.include?(:generic)
+
+            puts "Building the TEST stack"
+
+            knife_cmd = Chef::Knife::StackBuild.new
+            knife_cmd.name_args = [ @test_data_path + '/test_repo/stack2.yml' ]
+            knife_cmd.config[:stack_id] = @stack_ids[:test]
+            knife_cmd.config[:environment] = 'TEST'
+            knife_cmd.config[:repo_path] = @repo_path
+            knife_cmd.run
+
+            knife_cmd = Chef::Knife::NodeList.new
+            node_list = run_knife(knife_cmd).split
+            expect(node_list).to include('test-TEST-0')
+
+            knife_cmd = Chef::Knife::Ssh.new
+            knife_cmd.name_args = [ 'name:test-TEST-0', "sudo su -c '[ -e ~/stack_configured ] && echo yes'" ]
+            knife_cmd.config[:attribute] = 'ipaddress'
+            knife_cmd.config[:ssh_user] = 'vagrant'
+            knife_cmd.config[:identity_file] = Dir.home + '/.vagrant/insecure_key'
+            result = run_knife(knife_cmd).chomp
+            expect(result[/\d+\.\d+\.\d+\.\d+ (\w+)/, 1]).to eq('yes')
+
+            knife_cmd = Chef::Knife::Ssh.new
+            knife_cmd.name_args = [ 'name:test-TEST-0', "sudo su -c 'cat /etc/chef/encrypted_data_bag_secret'" ]
+            knife_cmd.config[:attribute] = 'ipaddress'
+            knife_cmd.config[:ssh_user] = 'vagrant'
+            knife_cmd.config[:identity_file] = Dir.home + '/.vagrant/insecure_key'
+            result = run_knife(knife_cmd).chomp
+
+            expect(result[/\d+\.\d+\.\d+\.\d+ (.*)/, 1]).to eq(IO.read("#{@repo_path}/secrets/TEST"))
+
+            puts "Destroying the TEST stack"
+
+            stack = StackBuilder::Stack::Stack.new(
+                StackBuilder::Chef::NodeProvider.new(@repo_path, 'TEST'),
+                @test_data_path + '/test_repo/stack2.yml',
+                @stack_ids[:test] )
+
+            stack.destroy
+
+            knife_cmd = Chef::Knife::NodeList.new
+            node_list = run_knife(knife_cmd).split
+            expect(node_list).to_not include('test-TEST-0')
+
+            knife_cmd = Chef::Knife::VagrantServerList.new
+            knife_cmd.config[:vagrant_dir] = File.join(Dir.home, '/.vagrant')
+            server_list = run_knife(knife_cmd)
+            expect(server_list.lines.keep_if { |l| l=~/test-TEST-0/ }.empty?).to be_truthy
+        end
+
+        if RUN_TESTS.include?(:containter)
+
+            puts "Creating a container stack"
+
+            knife_cmd = Chef::Knife::StackBuild.new
+            knife_cmd.name_args = [ @test_data_path + '/test_repo/stack3.yml' ]
+            knife_cmd.config[:stack_id] = @stack_ids[:dev]
+            knife_cmd.config[:environment] = 'DEV'
+            knife_cmd.config[:repo_path] = @repo_path
+            knife_cmd.run
+
+            node = Chef::Node.load('database-DEV-0')
+            container_port_map = node['container_port_map']
+
+            expect(container_port_map['wordpress'].nil?).to be_falsey
+            expect(container_port_map['wordpress'].size).to eq(1)
+
+            validate_nodes(['database-DEV-0', 'wordpress_web-DEV-0'], node)
+
+            puts "Scaling a container stack"
+
+            knife_cmd = Chef::Knife::StackBuild.new
+            knife_cmd.name_args = [ @test_data_path + '/test_repo/stack3.yml' ]
+            knife_cmd.config[:stack_id] = @stack_ids[:dev]
+            knife_cmd.config[:node] = 'wordpress_web:3'
+            knife_cmd.config[:environment] = 'DEV'
+            knife_cmd.config[:repo_path] = @repo_path
+            knife_cmd.run
+
+            node = Chef::Node.load('database-DEV-0')
+            container_port_map = node['container_port_map']
+            expect(container_port_map['wordpress'].size).to eq(3)
+
+            validate_nodes([ 'database-DEV-0', 'wordpress_web-DEV-0',
+                'wordpress_web-DEV-1', 'wordpress_web-DEV-2' ], node)
+
+            knife_cmd = Chef::Knife::StackBuild.new
+            knife_cmd.name_args = [ @test_data_path + '/test_repo/stack3.yml' ]
+            knife_cmd.config[:stack_id] = @stack_ids[:dev]
+            knife_cmd.config[:node] = 'wordpress_web:1'
+            knife_cmd.config[:environment] = 'DEV'
+            knife_cmd.config[:repo_path] = @repo_path
+            knife_cmd.run
+
+            node = Chef::Node.load('database-DEV-0')
+            container_port_map = node['container_port_map']
+            expect(container_port_map['wordpress'].size).to eq(1)
+
+            validate_nodes(['database-DEV-0', 'wordpress_web-DEV-0', ], node)
+
+            puts "Destroying a container stack"
+
+            stack = StackBuilder::Stack::Stack.new(
+                StackBuilder::Chef::NodeProvider.new(@repo_path, 'DEV'),
+                @test_data_path + '/test_repo/stack3.yml',
+                @stack_ids[:dev] )
+
+            stack.destroy
+        end
     end
-
-    it "should scale down the DEV stack web tier" do
-
-        knife_cmd = Chef::Knife::StackBuild.new
-        knife_cmd.name_args = [ @test_data_path + '/test_repo/stack1.yml' ]
-        knife_cmd.config[:stack_id] = @stack_ids[:dev]
-        knife_cmd.config[:environment] = 'DEV'
-        knife_cmd.config[:node] = 'wordpress:1'
-        knife_cmd.config[:repo_path] = @repo_path
-        knife_cmd.run
-
-        validate_nodes(['database-DEV-0', 'wordpress-DEV-0', 'loadbalancer-DEV-0'])
-    end
-
-    it "should destroy the DEV stack" do
-
-        stack = StackBuilder::Stack::Stack.new(
-            StackBuilder::Chef::NodeProvider.new(@repo_path, 'DEV'),
-            @test_data_path + '/test_repo/stack1.yml',
-            @stack_ids[:dev] )
-
-        stack.destroy
-
-        knife_cmd = Chef::Knife::NodeList.new
-        node_list = run_knife(knife_cmd).split
-        expect(node_list.empty?).to be_truthy
-    end
-
-    it "should build TEST stack" do
-
-        knife_cmd = Chef::Knife::StackBuild.new
-        knife_cmd.name_args = [ @test_data_path + '/test_repo/stack2.yml' ]
-        knife_cmd.config[:stack_id] = @stack_ids[:test]
-        knife_cmd.config[:environment] = 'TEST'
-        knife_cmd.config[:repo_path] = @repo_path
-        knife_cmd.run
-
-        knife_cmd = Chef::Knife::NodeList.new
-        node_list = run_knife(knife_cmd).split
-        expect(node_list).to include('test-TEST-0')
-
-        knife_cmd = Chef::Knife::Ssh.new
-        knife_cmd.name_args = [ 'name:test-TEST-0', "sudo su -c '[ -e ~/stack_configured ] && echo yes'" ]
-        knife_cmd.config[:attribute] = 'ipaddress'
-        knife_cmd.config[:ssh_user] = 'vagrant'
-        knife_cmd.config[:identity_file] = Dir.home + '/.vagrant/insecure_key'
-        result = run_knife(knife_cmd).chomp
-        expect(result[/\d+\.\d+\.\d+\.\d+ (\w+)/, 1]).to eq('yes')
-
-        knife_cmd = Chef::Knife::Ssh.new
-        knife_cmd.name_args = [ 'name:test-TEST-0', "sudo su -c 'cat /etc/chef/encrypted_data_bag_secret'" ]
-        knife_cmd.config[:attribute] = 'ipaddress'
-        knife_cmd.config[:ssh_user] = 'vagrant'
-        knife_cmd.config[:identity_file] = Dir.home + '/.vagrant/insecure_key'
-        result = run_knife(knife_cmd).chomp
-
-        expect(result[/\d+\.\d+\.\d+\.\d+ (.*)/, 1]).to eq(IO.read("#{@repo_path}/secrets/TEST"))
-    end
-
-    it "should destroy TEST stack" do
-
-        stack = StackBuilder::Stack::Stack.new(
-            StackBuilder::Chef::NodeProvider.new(@repo_path, 'TEST'),
-            @test_data_path + '/test_repo/stack2.yml',
-            @stack_ids[:test] )
-
-        stack.destroy
-
-        knife_cmd = Chef::Knife::NodeList.new
-        node_list = run_knife(knife_cmd).split
-        expect(node_list).to_not include('test-TEST-0')
-
-        knife_cmd = Chef::Knife::VagrantServerList.new
-        knife_cmd.config[:vagrant_dir] = File.join(Dir.home, '/.vagrant')
-        server_list = run_knife(knife_cmd)
-        expect(server_list.lines.keep_if { |l| l=~/test-TEST-0/ }.empty?).to be_truthy
-    end
-
-    it "should create a container stack" do
-
-        knife_cmd = Chef::Knife::StackBuild.new
-        knife_cmd.name_args = [ @test_data_path + '/test_repo/stack3.yml' ]
-        knife_cmd.config[:stack_id] = @stack_ids[:dev]
-        knife_cmd.config[:environment] = 'DEV'
-        knife_cmd.config[:repo_path] = @repo_path
-        knife_cmd.run
-
-        node = Chef::Node.load('database-DEV-0')
-        container_port_map = node['container_port_map']
-
-        expect(container_port_map['wordpress'].nil?).to be_falsey
-        expect(container_port_map['wordpress'].size).to eq(1)
-
-        validate_nodes(['database-DEV-0', 'wordpress_web-DEV-0'], node)
-
-        knife_cmd = Chef::Knife::StackBuild.new
-        knife_cmd.name_args = [ @test_data_path + '/test_repo/stack3.yml' ]
-        knife_cmd.config[:stack_id] = @stack_ids[:dev]
-        knife_cmd.config[:node] = 'wordpress_web:3'
-        knife_cmd.config[:environment] = 'DEV'
-        knife_cmd.config[:repo_path] = @repo_path
-        knife_cmd.run
-
-        node = Chef::Node.load('database-DEV-0')
-        container_port_map = node['container_port_map']
-        expect(container_port_map['wordpress'].size).to eq(3)
-
-        validate_nodes([ 'database-DEV-0', 'wordpress_web-DEV-0',
-            'wordpress_web-DEV-1', 'wordpress_web-DEV-2' ], node)
-
-        knife_cmd = Chef::Knife::StackBuild.new
-        knife_cmd.name_args = [ @test_data_path + '/test_repo/stack3.yml' ]
-        knife_cmd.config[:stack_id] = @stack_ids[:dev]
-        knife_cmd.config[:node] = 'wordpress_web:1'
-        knife_cmd.config[:environment] = 'DEV'
-        knife_cmd.config[:repo_path] = @repo_path
-        knife_cmd.run
-
-        node = Chef::Node.load('database-DEV-0')
-        container_port_map = node['container_port_map']
-        expect(container_port_map['wordpress'].size).to eq(1)
-
-        validate_nodes(['database-DEV-0', 'wordpress_web-DEV-0', ], node)
-
-        stack = StackBuilder::Stack::Stack.new(
-            StackBuilder::Chef::NodeProvider.new(@repo_path, 'DEV'),
-            @test_data_path + '/test_repo/stack3.yml',
-            @stack_ids[:dev] )
-
-        stack.destroy
-    end
-
 end
